+++ changelog.md
@@ -1,2 +1,7 @@
+# changelog  (25/10/2018)
+- `Core\Config` deleted old functions
+- `Core\Debug` deleted old functions
+- `Admin::removeAccentsToUploadFiles()` and `Admin::removeFrenchPonctuationToUploadFiles()` deleted, we now use `wpackagist-plugin/bea-sanitize-filename` package
+
 # changelog  (31/07/2018)
 -  deprecated `Admin::removeAccentsToUploadFiles()` and `Admin::removeFrenchPonctuationToUploadFiles()`, we now use "bea/sanitize-filename" plugin